+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of lizard-ui
 3.7 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Added ``live: true`` to tipsy tooltips so that elements created later can also get tooltips
 
 
 3.6 (2011-10-28)